+++ cmd/checkfleet/main.go
@@ -29,6 +29,7 @@ import (
 	"github.com/Allan-Nava/checkfleet/internal/checks/nats"
 	"github.com/Allan-Nava/checkfleet/internal/checks/patroni"
 	"github.com/Allan-Nava/checkfleet/internal/checks/postgres"
+	"github.com/Allan-Nava/checkfleet/internal/checks/redis"
 	"github.com/Allan-Nava/checkfleet/internal/checks/stream"
 	"github.com/Allan-Nava/checkfleet/internal/engine"
 	"github.com/Allan-Nava/checkfleet/internal/history"
@@ -73,7 +74,7 @@ func main() {
 
 func usage() {
 	fmt.Fprintln(os.Stderr, `uso:
-  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul|postgres|dns> --config checkfleet.yml [--output text|markdown|json|slack] [--only ...] [--min-severity warn] [--target glob] [--exit-on-bad]
+  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul|postgres|dns|redis> --config checkfleet.yml [--output text|markdown|json|slack] [--only ...] [--min-severity warn] [--target glob] [--exit-on-bad]
   checkfleet serve --config checkfleet.yml [--listen :9876] [--interval 60s]   # esporta le metriche Prometheus
   checkfleet report-issues --config checkfleet.yml [--dry-run]                 # apre/chiude issue GitHub dai finding BAD
   checkfleet validate --config checkfleet.yml                                  # valida la config senza eseguire i check
@@ -288,6 +289,7 @@ func modules(cfg *engine.Config) []moduleSpec {
 		{"consul", c.Consul != nil, func() engine.Check { return consul.New(*c.Consul) }},
 		{"postgres", c.Postgres != nil, func() engine.Check { return postgres.New(*c.Postgres) }},
 		{"dns", c.DNS != nil, func() engine.Check { return dns.New(*c.DNS) }},
+		{"redis", c.Redis != nil, func() engine.Check { return redis.New(*c.Redis) }},
 	}
 }